requirements.txt
streamlit
python-dotenv
sentence-transformers
faiss-cpu
pdfplumber
numpy
tqdm
openai>=1.0.0
requests
